+++ projects/webui/channelconfigurator/src/main/resources/js/devices/devicesController.js
@@ -7,8 +7,11 @@
 		$translate('DEVICE_DELETED_SUCCESSFULLY').then(function(text) {
 			$scope.deviceOKText = text;
 		});
-		
-		
+
+        $translate('DELETE_CONFIRM_MESSAGE').then(function(confirmMessage) {
+            $scope.confirmMessage = confirmMessage;
+        });
+
 		$scope.drivers = [];
 		
 		DriversService.getDrivers().then(function(drivers){
@@ -22,16 +25,18 @@
 		});
 
 		$scope.deleteDevice = function(deviceName) {
-			DevicesService.destroy(deviceName).then(function(data) {
-				$alert({content: $scope.deviceOKText, type: 'success'});
-				return $state.go('channelconfigurator.devices.index');
-			});
-
-			$.each($scope.drivers, function(index, driver) {
-				DevicesService.getDevices(driver).then(function(devices){
-					$scope.drivers[index]['devices'] = devices;
+            if (confirm($scope.confirmMessage + " " + deviceName + "?") == true) {
+				DevicesService.destroy(deviceName).then(function(data) {
+					$alert({content: $scope.deviceOKText, type: 'success'});
+					return $state.go('channelconfigurator.devices.index');
 				});
-			});
+
+				$.each($scope.drivers, function(index, driver) {
+					DevicesService.getDevices(driver).then(function(devices){
+						$scope.drivers[index]['devices'] = devices;
+					});
+				});
+			}
 		};
 		
 	};
